move plugin to its own file
