+++ src/mpibackend4jax/__init__.py
@@ -6,30 +6,11 @@ by setting the required environment variables when imported.
 """
 
 import os
-from pathlib import Path
 
 # Import the cluster to register it automatically
 # from .mpitrampoline_cluster import MPITrampolineLocalCluster
 
 __version__ = "0.1.0"
-
-def initialize():
-    # Get the package installation directory
-    _package_dir = Path(__file__).parent
-    _mpiwrapper_lib = _package_dir / "lib" / "libmpiwrapper.so"
-
-    # Set environment variables for MPITrampoline
-    if _mpiwrapper_lib.exists():
-        if "MPITRAMPOLINE_LIB" not in os.environ.keys():
-            os.environ["MPITRAMPOLINE_LIB"] = str(_mpiwrapper_lib.absolute())
-            print(f"mpibackend4jax: Set MPITRAMPOLINE_LIB={_mpiwrapper_lib.absolute()}")
-        if "JAX_CPU_COLLECTIVES_IMPLEMENTATION" not in os.environ.keys()
-            os.environ["JAX_CPU_COLLECTIVES_IMPLEMENTATION"] = "mpi"
-            print("mpibackend4jax: Set JAX_CPU_COLLECTIVES_IMPLEMENTATION=mpi")
-    else:
-        print(f"Warning: MPIWrapper library not found at {_mpiwrapper_lib}")
-        print("Please ensure the package was installed correctly.")
-
 
 # # Convenience function to check if MPITrampoline is properly configured
 # def is_configured():
